jbd2: fix sending of data flush on journal commit
In data=ordered mode, it's theoretically possible (however rare) that
an inode is filed to transaction's t_inode_list and a flusher thread
writes all the data and inode is reclaimed before the transaction
starts to commit.  In such a case, we could erroneously omit sending a
flush to file system device when it is different from the journal
device (because data can still be in disk cache only).

Fix the problem by setting a flag in a transaction when some inode is added
to it and then send disk flush in the commit code when the flag is set.

[Note from Zheng Liu]
This patch has been backported into the latest rhel kernel (358).  So drop
it after rebasing against the latest kernel.

Reference: 81be12c8179c1c397d3f179cdd9b3f7146cf47f1
Patch-in-mainline: v2.6.39-rc3-43-g81be12c8179c
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Zheng Liu <wenqing.lz@taobao.com>